organize endpoints and reconfigure security
