+++ app/src/main/resources/application-local.properties
@@ -1,0 +1,1 @@
+storage.folder=../local/files/
