+++ app/src/main/resources/application.properties
@@ -17,7 +17,6 @@ logging.file=accountrptgen.log
 build.version=@project.version@
 build.timestamp=@timestamp@
 
-
 spring.datasource.url=jdbc:h2:mem:test;DB_CLOSE_ON_EXIT=FALSE
 spring.datasource.username=test
 spring.datasource.password=test
