+++ app/src/test/java/net/ketone/accrptgen/it/AccrptgenApplicationITCase.java
@@ -1,5 +1,7 @@
-package net.ketone.accrptgen;
+package net.ketone.accrptgen.it;
 
+import net.ketone.accrptgen.dto.AccountFileDto;
+import net.ketone.accrptgen.gen.GenerationTest;
 import net.ketone.accrptgen.gen.Pipeline;
 import net.ketone.accrptgen.mail.Attachment;
 import net.ketone.accrptgen.mail.EmailService;
@@ -12,11 +14,17 @@ import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Profile;
+import org.springframework.test.annotation.IfProfileValue;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 
@@ -38,32 +46,26 @@ import static org.mockito.Matchers.eq;
  * which both contains sensitive information and is excluded
  * output is in target/test-output folder
  */
-// @Ignore
+@IfProfileValue(name = "spring.profiles.active", values = {"itcase"})
 @RunWith(SpringRunner.class)
 @ActiveProfiles("local")
 @SpringBootTest
-public class AccrptgenApplicationTests {
-
-	static private final String PLAIN_FILENAME = "program (plain) 09.4.18.xlsm";
-//	private final String COMPANY_NAME = "TestComp";
-	static private Date genTime = new Date();
-
-	@MockBean
-	private EmailService emailService;
-	@Autowired
-	private ApplicationContext ctx;
-
+public class AccrptgenApplicationITCase {
 
 	@TestConfiguration
 	static class AppConfig {
 
+		@Value("${plain.filename}${plain.filename.extension}")
+		private String PLAIN_FILENAME;
+
 		@Bean
+		@Primary
 		public StorageService storageService() throws IOException {
 			StorageService storageService = Mockito.mock(FileStorageService.class);
 
 			InputStream inputStream = this.getClass().getResourceAsStream("/" + PLAIN_FILENAME);
 
-			Mockito.when(storageService.load(eq(genTime.getTime()+".xlsm"))).thenReturn(IOUtils.toByteArray(inputStream));
+			Mockito.when(storageService.load(eq(PLAIN_FILENAME))).thenReturn(IOUtils.toByteArray(inputStream));
 			Mockito.when(storageService.loadAsInputStream(any(String.class))).thenCallRealMethod();
 
 			return storageService;
@@ -71,13 +73,28 @@ public class AccrptgenApplicationTests {
 
 	}
 
+
+	static private Date genTime = new Date();
+
+	@Value("${plain.filename}")
+	private String PLAIN_FILENAME;
+
+	@MockBean
+	private EmailService emailService;
+	@Autowired
+	private ApplicationContext ctx;
+
+
 	@Test
 	public void testPipeline() throws Exception {
 
 
 		ArgumentCaptor<List> argumentCaptor = ArgumentCaptor.forClass(List.class);
 
-		Pipeline pipeline = ctx.getBean(Pipeline.class, genTime);
+		AccountFileDto accountFileDto = new AccountFileDto();
+		accountFileDto.setGenerationTime(genTime);
+		accountFileDto.setFilename(PLAIN_FILENAME);
+		Pipeline pipeline = ctx.getBean(Pipeline.class, accountFileDto);
 		pipeline.run();
 		Mockito.verify(emailService).sendEmail(any(), argumentCaptor.capture());
 		List<Attachment> attachments = argumentCaptor.getValue();
