+++ app/src/test/resources/application-local.properties
@@ -1,0 +1,4 @@
+plain.filename=program (plain) 09.4.18
+plain.filename.extension=.xlsm
+template.filename=All documents 09.4.18-3.xlsm
+storage.folder=${user.dir}/../local/files/
